pyproject.toml
[tool.poetry]
name = "blog_uploader"
version = "0.1.0"
description = ""
authors = ["Zachary Juang <zachary822@me.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
lxml = "^4.8.0"
requests = "^2.27.1"
pymongo = "^4.1.1"
pydantic = "^1.9.0"
dnspython = "^2.2.1"
python-dotenv = "^0.20.0"
PyYAML = "^6.0"
Jinja2 = "^3.1.1"
boto3 = "^1.21.46"

[tool.poetry.dev-dependencies]
mypy = "^0.942"
types-PyYAML = "^6.0.7"
types-requests = "^2.27.25"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
